requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
colorama==0.4.4
coverage==5.5
Django==3.2.3
django-3-jet==1.0.8
django-admin-honeypot==1.1.0
django-debug-toolbar==3.2.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
pylint==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
mysqlclient==2.0.3
Pillow==8.0.1
